+++ src/components/tasks/sider/TaskSider.js
@@ -250,25 +250,35 @@ function TaskSider(props) {
                         key="general"
                         title={<Icon icon="home" text="General" />}
                         onTitleClick={({ key }) => onOpenChange(key)}>
-                        {getGeneralTaskFilters().map(taskFilter => (
+                        {getGeneralTaskFilters().filter(filter => settingsApi.settings['taskFilterVisible_' + filter.id] !== false).map(filter => (
                             <Menu.Item
-                                key={taskFilter.id}
-                                filter={taskFilter}>
+                                key={filter.id}
+                                filter={filter}>
                                 <LeftRight right={(
                                     <React.Fragment>
                                         <Icon
                                             icon="info-circle"
                                             color={Constants.fadeIconColor}
                                             className="object-actions"
-                                            onClick={() => setSelectedTaskFilterInfo(taskFilter)} />
-                                        {createBadge(taskFilter)}
+                                            onClick={() => setSelectedTaskFilterInfo(filter)} />
+                                        {createBadge(filter)}
                                     </React.Fragment>
                                 )}>
                                     <Icon
-                                        icon={taskFilter.icon}
-                                        color={taskFilter.color}
-                                        text={taskFilter.title} />
+                                        icon={filter.icon}
+                                        color={filter.color}
+                                        text={filter.title} />
                                 </LeftRight>
+                            </Menu.Item>
+                        ))}
+                        {taskFilterApi.taskFilters.filter(filter => filter.directory === 'general').map(filter => (
+                            <Menu.Item key={filter.id} filter={filter}>
+                                <ObjectMenuItem
+                                    badge={createBadge(filter)}
+                                    object={filter}
+                                    onManage={() => manageTaskFilters()}
+                                    onEdit={() => editTaskFilter(filter.id)}
+                                    onDelete={() => taskFilterApi.deleteTaskFilter(filter.id)} />
                             </Menu.Item>
                         ))}
                     </Menu.SubMenu>
@@ -378,14 +388,14 @@ function TaskSider(props) {
                     <Menu.SubMenu
                         key="taskFilters"
                         title={createCategorySubMenu('Task Filters', 'filter', () => manageTaskFilters(), () => onOpenChange('taskFilters'))}>
-                        {taskFilterApi.taskFilters.map(taskFilter => (
-                            <Menu.Item key={taskFilter.id} filter={taskFilter}>
+                        {taskFilterApi.taskFilters.filter(filter => filter.directory !== 'general').map(filter => (
+                            <Menu.Item key={filter.id} filter={filter}>
                                 <ObjectMenuItem
-                                    badge={createBadge(taskFilter)}
-                                    object={taskFilter}
+                                    badge={createBadge(filter)}
+                                    object={filter}
                                     onManage={() => manageTaskFilters()}
-                                    onEdit={() => editTaskFilter(taskFilter.id)}
-                                    onDelete={() => taskFilterApi.deleteTaskFilter(taskFilter.id)} />
+                                    onEdit={() => editTaskFilter(filter.id)}
+                                    onDelete={() => taskFilterApi.deleteTaskFilter(filter.id)} />
                             </Menu.Item>
                         ))}
                     </Menu.SubMenu>